refactor: move GameCard to MainPage folder
