+++ react-components/src/components/pages/MainPage/components/GameCard.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { IGameCard } from '../types';
+import { IGameCard } from '../../../../types';
 
 function GameCard(props: IGameCard): JSX.Element {
   return (
